Adding an example simulation test. (#63)
This change adds a test_nodes package with an example test node for evaluating robot behavior in simulation. The test node asserts that the robot can navigate to its goal successfully. The test node uses the Tag and Cancel hosted-ros-node services to record test status and self-terminate. New launch files are added for running the robot code in the small house world. The number of successful navigation to check for and the simulation world to run in can be configured with environment variables. If the environment variables are not set the defaults are '1' and 'bookstore'. The launch file and code exercised by the "1-click" workflow are left unchanged.

These changes have been tested by invoking the test_world launch file in AWS RoboMaker simulation to trigger 4 tests:

    a single navigation in the bookstore world
    4 navigations in the bookstore world
    a single navigation in the small house world
    4 navigations in the small house world