Corrected call to node test_nodes in test_world.launch (#153)
When launching test_nodes.launch after following the steps in the original repo, an error would happen in which the test node doesn't recognize "navigation_test". This is because it's missing the .py at the end. Its a simple fix which fixed it for me. 
Without this correction, the world and robot simulation would work, but not the test node.